修复主要编译错误
修复的问题：
1. ServiceRegistration.cs - 添加Offline命名空间前缀
2. GPUResourceManager.cs - 修复AllocateGPUAsync重载调用
3. RealtimePipelineService.cs - 修复GPUWorkloadType命名空间和枚举值
4. RealtimePipelineService.cs - 修复VoiceId改为Voice
5. RealtimePipelineService.cs - 注释掉未实现的GenerateRealtimeVideoAsync
6. RealtimePipelineService.cs - 修复sessionId未定义错误
7. AudioPipelineService.cs - 修复TimeSpan.FromSeconds类型转换
8. StreamingOllamaService.cs - 修复PostAsync参数

还有一些Streaming服务的错误需要后续修复

+++ LmyDigitalHuman/Services/Core/AudioPipelineService.cs
@@ -185,7 +185,7 @@ namespace LmyDigitalHuman.Services.Core
                         Success = true,
                         Text = whisperResult.Text,
                         Confidence = 0.95f, // Whisper通常有较高的准确率
-                        Duration = TimeSpan.FromSeconds(whisperResult.Duration),
+                        Duration = TimeSpan.FromSeconds((double)whisperResult.Duration),
                         ProcessingTime = stopwatch.ElapsedMilliseconds,
                         Language = "zh-CN",
                         Segments = ParseWhisperSegments(whisperResult.Text)
